+++ puma/tests/utils/test_confusion_matrix.py
@@ -14,7 +14,7 @@ class ConfusionMatrixTestCase(unittest.TestCase):
     def test_confusion_matrix_unweighted(self):
         targets = np.array([2, 0, 2, 2, 0, 1])
         predictions = np.array([0, 0, 2, 2, 0, 2])
-        cm = confusion_matrix(targets, predictions, normalize="all")
+        cm, _, _ = confusion_matrix(targets, predictions, normalize="all")
         expected_cm = np.array([
             [0.33333333, 0.0, 0.0],
             [0.0, 0.0, 0.16666667],
@@ -25,14 +25,14 @@ class ConfusionMatrixTestCase(unittest.TestCase):
     def test_confusion_matrix_unweighted_colnorm(self):
         targets = np.array([2, 0, 2, 2, 0, 1])
         predictions = np.array([0, 0, 2, 2, 0, 1])
-        cm = confusion_matrix(targets, predictions, normalize="colnorm")
+        cm, _, _ = confusion_matrix(targets, predictions, normalize="colnorm")
         expected_cm = np.array([[0.66666667, 0.0, 0.0], [0.0, 1.0, 0.0], [0.33333333, 0.0, 1.0]])
         np.testing.assert_array_almost_equal(expected_cm, cm)
 
     def test_confusion_matrix_unweighted_rownorm(self):
         targets = np.array([2, 0, 2, 2, 0, 1])
         predictions = np.array([0, 0, 2, 2, 0, 1])
-        cm = confusion_matrix(targets, predictions, normalize="rownorm")
+        cm, _, _ = confusion_matrix(targets, predictions, normalize="rownorm")
         expected_cm = np.array([
             [1.0, 0.0, 0.0],
             [0.0, 1.0, 0.0],
@@ -43,15 +43,19 @@ class ConfusionMatrixTestCase(unittest.TestCase):
     def test_confusion_matrix_unweighted_normNone(self):
         targets = np.array([2, 0, 2, 2, 0, 1])
         predictions = np.array([0, 0, 2, 2, 0, 1])
-        cm = confusion_matrix(targets, predictions, normalize=None)
+        cm, eff, fake = confusion_matrix(targets, predictions, normalize=None)
         expected_cm = np.array([[2.0, 0.0, 0.0], [0.0, 1.0, 0.0], [1.0, 0.0, 2.0]])
+        expected_eff = np.array([1.0, 1.0, 0.66666667])
+        expected_fake = np.array([0.0, 0.0, 0.33333333])
         np.testing.assert_array_almost_equal(expected_cm, cm)
+        np.testing.assert_array_almost_equal(expected_eff, eff)
+        np.testing.assert_array_almost_equal(expected_fake, fake)
 
     def test_confusion_matrix_weighted(self):
         targets = np.array([2, 0, 2, 2, 0, 1])
         predictions = np.array([0, 0, 2, 2, 0, 2])
         weights = np.array([1, 0.5, 0.5, 1, 0.2, 1])
-        cm = confusion_matrix(targets, predictions, sample_weights=weights, normalize="all")
+        cm, _, _ = confusion_matrix(targets, predictions, sample_weights=weights, normalize="all")
         expected_cm = np.array([
             [0.16666667, 0.0, 0.0],
             [0.0, 0.0, 0.23809524],